feat: переработан расчет оценки
